refactor(data-classes): AppSync location
