+++ aws_lambda_powertools/utilities/data_classes/appsync/appsync_resolver_event.py
